backend/requirements.txt
asgiref==3.8.1
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.3
defusedxml==0.8.0rc2
Django==3.2.16
django-cors-headers==3.13.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.10
gunicorn==20.1.0
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
oauthlib==3.3.1
packaging==25.0
Pillow==9.5.0
psycopg2-binary==2.9.5
pycparser==2.23
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
drf-yasg==1.21.4